# Add support to the small reduction transform to statically split small vectors incrementally by 4.