This brings nice improvements to small sizes divisible by 4.
  
Before this PR:
```
With transform dialect: reduction_2d_static --function_input="1234567x44xf32" P50: 602210.00 ns 90.20266684379203267963 GElements/s
Without transform dialect: reduction_2d_static --function_input="1234567x44xf32" P50: 2444700.0000 ns 22.21988301223053953450 GElements/s
```

With this PR:
```
With transform dialect: reduction_2d_static --function_input="1234567x44xf32" P50: 417270.00 ns 130.18177199415246722745 GElements/s
Without transform dialect: reduction_2d_static --function_input="1234567x44xf32" P50: 2221600.0000 ns 24.45127295642779978393 GElements/s
```

For non-divisible and dynamic sizes, one needs an adaptive split that adapts to the future stride + offset of the tensor to ensure dynamic alignemnt.
This is left for future work.
The performance drop on non-naturally aligned sizes is visible with e.g.:

```
(TRANSFORM_DIALECT_NO_DEBUG=1  benchmark-transform-create -r  ../iree-samples/transform_dialect/benchmark_linalg_reductions.stub.mlir   reduction_2d_static 1234567 39 )
```

Printing:

```
With transform dialect: reduction_2d_static --function_input="1234567x39xf32" P50: 497440.00 ns 96.79180001608234158893 GElements/s
Without transform dialect: reduction_2d_static --function_input="1234567x39xf32" P50: 10705000.000 ns 4.49772190565156468939 GElements/s
```