@nicolasvasilache 2022-12-30
> Could be nice to have a test for the emitted transform sequence.
> 
> Out of curiosity, could we achieve the same effect by splitting once after `4 * ((strategy.captures.reductionDimensionSize - 1) / 4)` and tiling the first part by 4?

Yes, but caveats .. I added a repro of an issue related to stuff @matthias-springer  has been touching https://github.com/iree-org/iree/pull/11681